backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
SQLAlchemy==2.0.36
pydantic==2.10.2
PyJWT==2.10.1
